+++ act/nari/ext/act/actlog.py
@@ -7,7 +7,7 @@ from nari.types.event.version import SemanticVersion, Version
 from nari.util.exceptions import EventNotFound
 
 from nari.ext.act.actlogutils import ID_MAPPINGS, ActEventType, ActLogChecksumType, date_from_act_timestamp, validate_checksum
-from nari.ext.act.actlogutils.exceptions import InvalidActChecksum
+from nari.ext.act.exceptions import InvalidActChecksum
 
 # The last version of ACT to use MD5 checksums
 # We default to SHA256, but if the version is <= 2.2.1.6 we switch to MD5
